requirements.txt
flask 
google-generativeai
pypdf